requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.6
django-cors-headers==4.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
openpyxl==3.1.2
Pillow==10.0.1
python-barcode==0.15.1
pytz==2023.3.post1
reportlab==4.0.5
sqlparse==0.4.4
typing-extensions==4.8.0
xlwt==1.3.0
